Normalize coding style

+++ PhpManager/private/Get-FileFromUrlOrCache.ps1
@@ -1,4 +1,4 @@
-Function Get-FileFromUrlOrCache
+function Get-FileFromUrlOrCache
 {
     <#
     .Synopsis
@@ -20,7 +20,7 @@ Function Get-FileFromUrlOrCache
     [System.Array]
     #>
     [OutputType([string], [bool])]
-    Param (
+    param (
         [Parameter(Mandatory = $True, Position = 0)]
         [ValidateNotNull()]
         [ValidateLength(1, [int]::MaxValue)]
@@ -28,92 +28,89 @@ Function Get-FileFromUrlOrCache
         [Parameter(Mandatory = $False, Position = 1)]
         [string] $CachedFileName
     )
-    Begin {
+    begin {
         $localFile = $null
         $fromCache = $null
     }
-    Process {
+    process {
         $extension = ''
-        If ($null -ne $CachedFileName -and $CachedFileName -ne '') {
+        if ($null -ne $CachedFileName -and $CachedFileName -ne '') {
             $match = $CachedFileName | Select-String -Pattern '.(\.[A-Za-z0-9_\-]+)$'
-            If ($match) {
+            if ($match) {
                 $extension = $match.Matches.Groups[1].Value
             }
-        } Else {
+        } else {
             $CachedFileName = ''
             $match = $Url | Select-String -Pattern '^[^:]+:/+[^?#/]+/(?:[^?#/]*/)*([^?#/]+)(?:$|^|#)'
-            If ($match) {
+            if ($match) {
                 $nameFromUrl = $match.Matches.Groups[1].Value
                 $match = $nameFromUrl | Select-String -Pattern '.(\.[A-Za-z0-9_\-]+)$'
-                If ($match) {
+                if ($match) {
                     $extension = $match.Matches.Groups[1].Value
                 }
-                If ($nameFromUrl -imatch '^[a-z0-9_\-][a-z0-9_\-\.]*[a-z0-9_\-]$') {
+                if ($nameFromUrl -imatch '^[a-z0-9_\-][a-z0-9_\-\.]*[a-z0-9_\-]$') {
                     $CachedFileName = $nameFromUrl;
                 }
             }
-            If ($CachedFileName -eq '') {
+            if ($CachedFileName -eq '') {
                 $stream = New-Object System.IO.MemoryStream
-                Try {
+                try {
                     $streamWriter = New-Object -TypeName System.IO.BinaryWriter -ArgumentList @($stream)
-                    Try {
+                    try {
                         $streamWriter.Write([System.Text.Encoding]::UTF8.GetBytes($Url))
                         $streamWriter.Flush()
                         $stream.Position = 0
                         $hash = Get-FileHash -InputStream $stream -Algorithm SHA1
                         $CachedFileName = $hash.Hash
-                    }
-                    Finally {
+                    } finally {
                         $streamWriter.Dispose()
                     }
-                }
-                Finally {
+                } finally {
                     $stream.Dispose()
                 }
             }
         }
         $downloadCachePath = Get-PhpDownloadCache
-        If ($downloadCachePath -eq '') {
+        if ($downloadCachePath -eq '') {
             $fullCachePath = ''
-        } Else {
-            If (-Not(Test-Path -LiteralPath $downloadCachePath -PathType Container)) {
+        } else {
+            if (-Not(Test-Path -LiteralPath $downloadCachePath -PathType Container)) {
                 New-Item -Path $downloadCachePath -ItemType Directory | Out-Null
             }
             $fullCachePath = Join-Path -Path $downloadCachePath -ChildPath $CachedFileName
         }
-        If ($fullCachePath -ne '' -and (Test-Path -LiteralPath $fullCachePath -PathType Leaf)) {
+        if ($fullCachePath -ne '' -and (Test-Path -LiteralPath $fullCachePath -PathType Leaf)) {
             $localFile = $fullCachePath
             $fromCache = $true
-        } Else {
+        } else {
             $temporaryFile = Get-TemporaryFileWithExtension -Extension $extension
-            Try {
-                Try {
+            try {
+                try {
                     [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12 + [Net.SecurityProtocolType]::Tls11 + [Net.SecurityProtocolType]::Tls
-                }
-                Catch {
+                } catch {
                     Write-Debug '[Net.ServicePointManager] or [Net.SecurityProtocolType] not found in current environment'
                 }
                 Write-Debug "Downloading from $Url"
                 Invoke-WebRequest -UseBasicParsing $Url -OutFile $temporaryFile
-                If ($fullCachePath -ne '') {
+                if ($fullCachePath -ne '') {
                     Move-Item -LiteralPath $temporaryFile -Destination $fullCachePath
                     $localFile = $fullCachePath
                     $fromCache = $true
-                } Else {
+                } else {
                     $localFile = $temporaryFile
                     $fromCache = $false
                 }
-            } Catch {
-                Try {
+            } catch {
+                try {
                     Remove-Item -LiteralPath $temporaryFile
-                } Catch {
+                } catch {
                     Write-Debug 'Failed to remove a temporary file'
                 }
-                Throw
+                throw
             }
         }
     }
-    End {
+    end {
         $localFile
         $fromCache
     }
